// src/pages/korean-agency/types.d.ts
export interface RegisterKoreanAgencyFormValues {
  email: string;
  password: string;
  agencyName: string;
  businessRegistrationNumber: string;
  representativeName: string;
  representativeMobilePhone: string;
  representativeBusinessPhone: string;
  representativeAddress: string;
  businessRegistrationCertificate: string;
}